msm: msm_fb: Remove broken code that attempted to align FB smem.
If passed an aligned buffer, the realignment in msm_fb_register would
offset the buffer by -4096 bytes, which caused the memory allocated
before the FB smem to be corrupted when the first 4096 bytes of the FB
were written.

This function is always passed an aligned buffer, so the useless
and broken realignment was removed, and a BUG_ON() added instead.

Change-Id: Iccc2da270ff882164a3fd7adc6a170b4c68e8fa4
Signed-off-by: Christopher Lais <chris+android@zenthought.org>